move all nocodb code to a separate optional package
