+++ src/packages/crm/README.md
@@ -1,0 +1,9 @@
+# @cocalc/crm
+
+Experimental backend support for CRM functionality will go here, maybe built on
+NocoDB. If this works out, we'll do a lot more in this direction. If not, we'll
+restart and write our own at a lower level.
+
+This package is NOT built by default. Unlike everything else in all of CoCalc,
+it is the one package where something it installs is AGPL licensed. Thus we
+avoid installing it by default and make its use optional.
