+++ src/packages/crm/src/index.ts
@@ -1,0 +1,2 @@
+import initNocoDB from "./nocodb";
+export { initNocoDB };
