+++ src/packages/hub/package.json
@@ -11,6 +11,7 @@
     "../assets",
     "../backend",
     "../cdn",
+    "../crm",
     "../database",
     "../frontend",
     "../next",
@@ -62,7 +63,6 @@
     "mime": "^1.3.4",
     "mkdirp": "^1.0.4",
     "next": "12.3.4",
-    "nocodb": "^0.100.2",
     "nodemailer": "^6.7.0",
     "nyc": "^15.1.0",
     "parse-domain": "^5.0.0",
@@ -87,7 +87,9 @@
     "uuid": "^8.3.2",
     "validator": "^13.6.0"
   },
+  "devDependenicesNotes": "For license and size reasons, we make @cocalc/crm a dev dependency so it is NOT installed unless explicitly installed as a separate step.",
   "devDependencies": {
+    "@cocalc/crm": "^1.0.0",
     "@types/http-proxy": "^1.17.9",
     "@types/jest": "^29.2.3",
     "@types/node": "^14.18.10",
